requirements.txt
flask>=2.0.0
flask-login>=0.6.0
flask-cors>=4.0.0
python-supabase>=2.0.0
werkzeug>=2.0.0
python-dotenv==0.19.0
python-jose==3.3.0
passlib==1.7.4 